pyproject.toml
[project]
name = "raguia-local-agent"
version = "0.2.0"
description = "Agent de synchronisation locale du dossier RAGUIA vers le portail Raguia"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "watchdog>=4.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
# Icone systray cross-platform (macOS, Windows, Linux)
tray = [
    "pystray>=0.19.0",
    "Pillow>=10.0.0",
]
# Toutes les fonctionnalites
full = [
    "raguia-local-agent[tray]",
]

[project.scripts]
raguia-local-agent = "raguia_local_agent.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["raguia_local_agent"]
